Create dbo.fnMatchText
Removes all vowels and non-alphabetic characters from a string (original function preserved numeric
characters, but that is commented out below - uncomment if useful) and converts all characters to
uppercase. Intended to be used before sending a string to the dbo.CompareText function to reduce
the number of comparisons it has to make and thereby increase efficiency.