Create DatabaseAdministration.fnDropCreateDefaultConstraintsWhereDefinitionContainsString
Returns a set of SQL statemets that allow the user to drop and then re-create all default constraints whose definition contains a specific string. Helpful for when you need to modify a function that is referenced in numerous default constraints as SQL Server prevents this without dropping and re-creating the constraints.